bikeshare/pyproject.toml
[project]
name = "bikeshare"
version = "0.1.0"
description = "Play with bikeshare data & snowflake"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.1.3",
    "pyyaml>=6.0.2",
    "snowflake-connector-python>=3.13.2",
]
